requirements.txt
fastapi
uvicorn

pymupdf
beautifulsoup4
requests

llama-index
sentence-transformers

qdrant-client

pillow
pytesseract
